Resolved soundexample linking problem in linux and removed a casting in kinectexample

+++ src/apps/kinectExample/main.cpp
@@ -103,7 +103,7 @@ void myPreSyncFun()
 		{
 			glm::vec3 leftPos = leftHand->getPosition();
 			glm::vec3 rightPos = rightHand->getPosition();
-			float dist = static_cast<float>(glm::length(leftPos - rightPos));
+			float dist = glm::length(leftPos - rightPos);
 			size_factor.setVal( (dist < 2.0f && dist > 0.2f) ? dist : 0.5f );
 		}
 	}
